re-work tests for new APE5 scheme
